+++ Cliente.cpp
@@ -4,7 +4,7 @@
 
 Cliente::Cliente(std::string nome,
                  std::string email,
-                 const int certificado,
+                 const long int certificado,
                  const int telefone,
                  bool tipo)
 : Pessoa(nome, email), _tipo(tipo), _telefone(telefone)
